backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
google-cloud-compute>=1.15.0
google-cloud-resource-manager>=1.12.0
pydantic>=2.5.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
google-cloud-container>=2.36.0
google-cloud-storage>=2.14.0
